requirements.txt
matplotlib
torch==1.9.1+cu111 torchvision==0.10.1+cu111 torchaudio==0.9.1 -f https://download.pytorch.org/whl/torch_stable.html
nibabel==5.2.0
numpy==1.24.4
pillow==10.2.0
monai==0.9.0
transformers==4.16.2
einops==0.3.2
pandas==1.4.1
psutil==5.8.0
lmdb==1.2.1
tqdm==4.62.3
gdown==4.2.1
tensorboard==2.6.0
scikit-image==0.19.1
itk==5.3.0 
pytorch-ignite==0.4.8
